// CaseStudy_CentralLibraryTracker/CentralLibraryTrackerFrontEnd/src/app/models/returnbook.ts
export class ReturnBook {
    id: number;
    bookId: number;
    memberId: number;
    subject: string;
    title: string;
    author: string;
    edition: number;
    pages: number;
    isbn: number;
    libraryName: string;
    name: string;
    email: string;
    major: string;
    issueDate: Date;
    expireDate: Date;
    returnDate: Date;
    status: number;
}